Show idle time based on input rather than output
Place additional file in the inprogressdir, same format but with
'.in' appended. User input is written to this file, so its modify time
updates when the user is playing.
Next, calculate idle_time based on the '.in' file instead of the ttyrec.
Add some checks to the code handling lockfiles in inprogressdir, to
ignore '.in' files that should be there and unlink orphaned ones.

squashed commits:
move input tracking file to inprogressdir
remove inputrec file on game exit
remove stale .in files